fixed z-equation and added more to CLI comments
