+++ app/mods/synth.py
@@ -1,420 +1,164 @@
-import argparse as ap
-import numpy as np
-import cv2
-import matplotlib
-from matplotlib import pyplot as plt
-from scipy.optimize import leastsq
+# libraries
+import argparse as ap                       # command line lib
+import paraboloid as pbd                    # paraboloid lib, paraboloid.py
+import numpy as np                          # numerical lib
+import matplotlib.pyplot as plt             # plotting lib
+import matplotlib.animation as animate      # animation lib
 
-matplotlib.use("Agg")
-
+# sets plots to a darker color scheme
+plt.style.use("dark_background")
 #------------------------------------------------------------------------------
 # *** COMMAND LINE INTERFACE SETUP ***
-
 # initial message for program
-msg = 'This program creates a synthetic .avi movie for use with DigiPyRo. '
-msg += 'The video shows a ball rolling on a parabolic surface. '
-msg += 'The user may change the length of the movie, the frame rate, the resolution of the movie, the frequency of oscillations, '
-msg += 'the rotation rate of the reference frame, and control the initial conditions of the ball.'
+msg = """ This program creates a synthetic .mp4 movie for use with DigiPyRo.
+The video shows a puck sliding on a parabolic surface. The user may change
+several variables of the puck, and observe the animation in the output."""
+
 parser = ap.ArgumentParser(description = msg,
                            formatter_class=ap.ArgumentDefaultsHelpFormatter)
 
 # collecting arguments for the user to change
 parser.add_argument('-t', '--time', type=float, default=5,
-                    help='The desired length of the movie in seconds.')
-parser.add_argument('-f', '--fps', type=float, default=30.0,
-                    help=('The frame rate of the video (frames per second). '
-                          'Set this to a low value (10-15) for increased speed'
-                          ' or a higher value (30-60) for better results.'))
-parser.add_argument('-w', '--width', type=int, default=1260,
-                    help='The width in pixels of the video.')
-parser.add_argument('-l', '--length', type=int, default=720,
-                    help='The height in pixels of the video.')
-parser.add_argument('-r', '--eqpot_rpm', type=float, default=10.0,
+                    help='The length of the movie in seconds.')
+
+parser.add_argument('-o', '--omega', type=float, default=3,
                     help=('The deflection of a stationary paraboloid surface '
-                          'as if it were an equipotentional in a system '
+                          'as if it were an equipotential in a system '
                           'rotating at the specified rate. '
                           'A good value would be between 5 and 15.'))
-parser.add_argument('-R', '--cam_rpm', type=float, default=0.0,
-                    help=('The rotation rate of the camera. The two natural '
-                          'frames of reference are with rotRate = 0 and '
-                          'rotRate = rpm.'))
-parser.add_argument('--r0', type=float, default=1.0,
-                    help=('The initial radial position of the ball. Choose a '
-                          'value betweeon 0 and 1.'))
-parser.add_argument('--vr0', type=float, default=0.0,
-                    help=('The initial radial velocity of the ball. '
-                          'A good value would be between 0 and 1.'))
-parser.add_argument('--phi0', type=float, default=np.pi/4,
-                    help=('The initial azimuthal position of the ball. '
-                          'Choose a value between 0 and 2*pi.'))
-parser.add_argument('--vphi0', type=float, default = 0,
-                    help=('The initial azimuthal velocity of the ball. '
-                          'A good value would be between 0 and 1.'))
+
+parser.add_argument('-u', '--u0', type=float, default=0,
+                    help=('The initial x-velocity of the puck moving on the '
+                          'parabolic surface. If too high, the puck will go '
+                          'off the edge during the animation.'))
+
+parser.add_argument('-v', '--v0', type=float, default=0,
+                    help=('The initial y-velocity of the puck moving on the '
+                          'parabolic surface. If too high, the puck will go '
+                          'off the edge during the animation.'))
+
+parser.add_argument('-x', '--x0', type=float, default=1,
+                    help=('The initial x-position of the puck moving on the '
+                          'parabolic surface. If too high, the puck will go '
+                          'off the edge during the animation.'))
+
+parser.add_argument('-r', '--radius', type=float, default=2,
+                    help='The radius of the paraboloid')
+
+parser.add_argument('-n', '--name', type=str, default="movie",
+                    help=("Title of the movie that will be outputted."
+                          "The extension does not need to be provided; the "
+                          ".mp4 will be added in the program."))
+
+parser.add_argument('-s', '--switch', type=int, default=0,
+                    help=('Choice to either show the animation or save it. '
+                          'Default displays the animation. Option [1] will '
+                          'save the animation as an mp4.'))
+#------------------------------------------------------------------------------
 
 # collecting user input into list
 args = parser.parse_args()
 
+time = args.time
+omega = args.omega
+u0 = args.u0
+v0 = args.v0
+x0 = args.x0
+radius = args.radius
+name = args.name + ".mp4"
+switch = args.switch
 
-spinlab = cv2.imread('util/SpinLabUCLA_BW_strokes.png') # spinlab logo to display in upper right corner of output video
+def animate_paraboloid(time, omega, u0, v0, x0, radius):
+    """Animates the paraboloid.
 
-
-# Define movie details
-movLength = args.time                   # [1] define the desired length of the movie in seconds
-fps = args.fps                          # [2] Set this to a low value (10-15) for increased speed or a higher value (30-60) for better results with DigiPyRo
-width = args.width                      # [3] Width and height in pixels
-height = args.length                    # [3] Decrease the width and height for increased speed, increase for improved resolution
-spinlab = cv2.resize(spinlab,(int(0.2*width),int((0.2*height)/3)), interpolation = cv2.INTER_CUBIC)
-
-
-# Define table values
-rpm = args.eqpot_rpm                    # [4] frequency of oscillations (in RPM). Good values might be 5-15
-                                        # NOTE: A two-dimensional rotating system naturally takes the shape of a parabola.
-                                        # The rotation rate determines the curvature of the parabola, which is why we define the curvature in terms of RPM
-rotRate = args.cam_rpm                  # [5] rotation rate of camera. The two natural frames of reference are with rotRate = 0 and rotRate = rpm
-
-
-# Set initial conditions
-r0 = args.r0                            # [6] initial radial position of ball. Choose a value between 0 and 1
-vr0 = args.vr0                          # [7] initial radial velocity of ball. Good values might be 0-1
-phi0 = args.phi0                        # [8] initial azimuthal position of ball. Choose a value between 0 and 2*pi
-vphi0 = args.vphi0                      # [9] initial azimuthal velocity of ball. Good values might be 0-1
-
-# Ask user for movie name
-saveFile = input('Enter a name for the movie (e.g. mySyntheticMovie): ')
-
-# Ask user if they would like a parabolic side view
-parView = input('Would you also like a side view? (yes/no): ')
-doParView = 'yes' in parView
-
-if doParView:
-    saveFilePar = saveFile + 'side.avi'
-saveFile += '.avi'
-
-if doParView:
-    parViewFrac = 0.3
-    borderHeight = 50
-    lineWidth = 10
-    parHeight = int(height*(parViewFrac+1))
-    fullHeight = parHeight + borderHeight
-else:
-    fullHeight = height
-
-
-# No longer asking user if they care about reference frame of parabolic side view
-doParViewRot = (rotRate != 0)
-
-# Ask user if they care about reference frame of parabolic side view
-#doParViewRot = False
-#if doParView and rotRate!=0:
-#    parViewRot = raw_input('Would you like the parabolic side view in the rotating frame? (yes/no): ')
-#    doParViewRot = 'yes' in parViewRot
-
-# Set the amplitude of oscillations to 40% of the smaller dimension
-amp = 0
-if width > height:
-    amp = int(0.5*height)
-    ballSize = int(height/30)
-else:
-    amp = int(0.5*width)
-    ballSize = int(width/30)
-
-omega = (rpm * 2 * np.pi)/60		# calculate angular frequency of oscillations
-
-# Create movie file
-fourcc = cv2.VideoWriter_fourcc('m','p','4','v')
-video_writer = cv2.VideoWriter(saveFile, fourcc, fps, (width, fullHeight))
-#if doParView:
-#    video_writer_par = cv2.VideoWriter(saveFilePar, fourcc, fps, (width, height))
-
-def r(t):
-    t1 = (((vr0**2)+((r0**2)*(vphi0**2)))*(np.sin(omega*t)**2))/(omega**2)
-    t2 = (1/omega)*(r0*vr0*np.sin(2*omega*t))
-    t3 = (r0**2)*(np.cos(omega*t)**2)
-    return (t1+2+t3)**(0.5)
-
-def phi(t):
-    y = ((1/omega)*(np.sin(omega*t))*(vr0*np.sin(phi0) + r0*vphi0*np.cos(phi0))) + r0*np.sin(phi0)*np.cos(omega*t)
-    x = ((1/omega)*(np.sin(omega*t))*(vr0*np.cos(phi0) - r0*vphi0*np.sin(phi0))) + r0*np.cos(phi0)*np.cos(omega*t)
-    return np.arctan2(y,x)
-
-def annotate(img, i, rotatingView): # puts diagnostic text info on each frame
-    font = cv2.FONT_HERSHEY_TRIPLEX
-
-    dpro = 'SynthPy'
-    dproLoc = (25, 50)
-    cv2.putText(img, dpro, dproLoc, font, 1, (255,255,255), 1)
-
-    topView = 'Top View'
-    topViewLoc = (25, 90)
-    cv2.putText(img, topView, topViewLoc, font, 1, (255,255,255), 1)
-
-    if rotatingView:
-        rotView = 'Rotating View'
-        rotViewLoc = (25, 130)
-        cv2.putText(img, rotView, rotViewLoc, font, 1, (55, 255, 90), 1)
-    else:
-        rotView = 'Inertial View'
-        rotViewLoc = (25, 130)
-        cv2.putText(img, rotView, rotViewLoc, font, 1, (255, 105, 180), 1)
-
-    img[25:25+spinlab.shape[0], (width-25)-spinlab.shape[1]:width-25] = spinlab
-
-    timestamp = 'Time: ' + str(round((i/fps),1)) + ' s'
-    tLoc = (width - 225, height-25)
-    cv2.putText(img, timestamp, tLoc, font, 1, (255, 255, 255), 1)
-
-    rad = 'Radius: R = 1 m'
-    radLoc = (width -325, height-65)
-    cv2.putText(img, rad, radLoc, font, 1, (255, 255, 255), 1)
-
-def parabolaPoints():
-    xpoints = np.empty(width)
-    ypoints = np.empty(width)
-    metersToPixels = float(amp)/2
-    for i in range(width):
-        if i < (width/2 - int(amp)) or i > (width/2 + int(amp)):
-            continue
-        xpoints[i] = i
-        #ypoints[i] = int( ((0.75)*float(fullHeight-height)) - ((omega**2)*((float(i-(width/2))/float(amp))*)/(2*9.8))
-        #ypoints[i] = int( ((0.75)*float(fullHeight-height)) - ((rpm**2)*((float(i-(width/2))/float(amp))**2))/(2*9.8))
-        ypoints[i] = int( ((0.75)*float(fullHeight-height)) - ((omega**2)*((float(i-(width/2))/float(amp))**2)*((metersToPixels)**2))/(2*9.8*metersToPixels))
-        nextPoint = np.array([xpoints[i], ypoints[i]])
-        try:
-            ppoints
-        except:
-            ppoints = nextPoint
-        else:
-            ppoints = np.append(ppoints, nextPoint, axis=0)
-    return ppoints
-
-def parabola(x):
-    metersToPixels = float(amp)/2
-    return int( ((0.75)*float(fullHeight-height)) - ((omega**2)*((float(x-(width/2))/float(amp))**2)*((metersToPixels)**2))/(2*9.8*metersToPixels))
-
-
-def createLineIterator(P1, P2, img):
-    """
-    Produces and array that consists of the coordinates and intensities of each pixel in a line between two points
-
-    Parameters:
-        -P1: a numpy array that consists of the coordinate of the first point (x,y)
-        -P2: a numpy array that consists of the coordinate of the second point (x,y)
-        -img: the image being processed
+    Keyword arguments:
+    time  -- float, length of animation in seconds
+    omega -- float, effective rotation
+    u0    -- float, initial x-component of the velocity
+    v0    -- float, initial y-component of the velocity
+    x0    -- float, initial x-component of the position
 
     Returns:
-        -it: a numpy array that consists of the coordinates and intensities of each pixel in the radii (shape: [numPixels, 3], row = [x,y,intensity])
+    animation -- Obj, animation object
+    NOTE: reference matplotlib.animation.funcAnimation for more info.
     """
-   #define local variables for readability
-    imageH = img.shape[0]
-    imageW = img.shape[1]
-    P1X = P1[0]
-    P1Y = P1[1]
-    P2X = P2[0]
-    P2Y = P2[1]
 
-   #difference and absolute difference between points
-   #used to calculate slope and relative location between points
-    dX = P2X - P1X
-    dY = P2Y - P1Y
-    dXa = np.abs(dX)
-    dYa = np.abs(dY)
+    # creating circle
+    circle = pbd.circle(radius)
+    parabola = pbd.parabola(radius, omega)
 
-   #predefine numpy array for output based on distance between points
-    itbuffer = np.empty(shape=(np.maximum(dYa,dXa),3))
-    itbuffer.fill(np.nan)
+    # size of plots based off paraboloid radius
+    rmax = radius*1.5
+    size = (-1*(rmax), rmax)
 
-   #Obtain coordinates along the line using a form of Bresenham's algorithm
-    negY = P1Y > P2Y
-    negX = P1X > P2X
-    if P1X == P2X: #vertical line segment
-        itbuffer[:,0] = P1X
-        if negY:
-            itbuffer[:,1] = np.arange(P1Y - 1,P1Y - dYa - 1,-1)
-        else:
-            itbuffer[:,1] = np.arange(P1Y+1,P1Y+dYa+1)
-    elif P1Y == P2Y: #horizontal line segment
-        itbuffer[:,1] = P1Y
-        if negX:
-            itbuffer[:,0] = np.arange(P1X-1,P1X-dXa-1,-1)
-        else:
-            itbuffer[:,0] = np.arange(P1X+1,P1X+dXa+1)
-    else: #diagonal line segment
-        steepSlope = dYa > dXa
-        if steepSlope:
-            slope = dX.astype(np.float64)/dY.astype(np.float64)
-            if negY:
-                itbuffer[:,1] = np.arange(P1Y-1,P1Y-dYa-1,-1)
-            else:
-                itbuffer[:,1] = np.arange(P1Y+1,P1Y+dYa+1)
-            itbuffer[:,0] = (slope*(itbuffer[:,1]-P1Y)).astype(np.int) + P1X
-        else:
-            slope = dY.astype(np.float64)/dX.astype(np.float64)
-            if negX:
-                itbuffer[:,0] = np.arange(P1X-1,P1X-dXa-1,-1)
-            else:
-                itbuffer[:,0] = np.arange(P1X+1,P1X+dXa+1)
-            itbuffer[:,1] = (slope*(itbuffer[:,0]-P1X)).astype(np.int) + P1Y
+    # creating figure and plots
+    fig, (a0, a1) = plt.subplots(2, 1, figsize=(6,8),
+                                 gridspec_kw={'height_ratios': [3,1]})
+    # a0 plots out the top-view of the paraboloid
+    a0.set_xlim(size)
+    a0.set_ylim(size)
+    a0.set_title("Top-View")
+    a0.set_xlabel("X")
+    a0.set_ylabel("Y", rotation=0)
 
-   #Remove points outside of image
-    colX = itbuffer[:,0]
-    colY = itbuffer[:,1]
-    itbuffer = itbuffer[(colX >= 0) & (colY >=0) & (colX<imageW) & (colY<imageH)]
+    a0.plot(circle[0], circle[1], color='white', label="Paraboloid")
+    puckTop, = a0.plot([], [], linestyle='none',
+                       marker='o', mfc='white', mec='red', label="Puck")
 
-   #Get intensities from img ndarray
-    itbuffer[:,2] = img[itbuffer[:,1].astype(np.uint),itbuffer[:,0].astype(np.uint)]
+    a0.grid(color='grey')
+    a0.legend()
 
-    return itbuffer
+    # a1 plots out the side-view of the paraboloid
+    a1.set_xlim(size)
+    a1.set_xlabel("X")
+    a1.set_ylabel("Z", rotation=0)
+    a1.set_title("Side-View")
 
+    a1.plot(parabola[0], parabola[1], color='white', label="Paraboloid")
+    puckSide, = a1.plot([], [], linestyle='none',
+                        marker='o', mfc='white', mec='red', label="Puck")
 
-def dottedLine(frame, xi, yi, xf, yf, c1, c2, c3, thickness, segmentLength):
-    it = createLineIterator((xi,yi), (xf, yf), frame[:,:,0])
-    totLength = (xf-xi)**2 + (yf-yi)**2
-    nLines = totLength//segmentLength
-    for i in range(nLines):
-        if i%2 == 0:
-            continue
-        #if (i)*segmentLength >= it.shape[0]:
-        #    continue
-        try:
-            cv2.line(frame, (it[i*segmentLength,0], it[i*segmentLength,1]), (it[(i+1)*segmentLength, 0], it[(i+1)*segmentLength, 1]), (c1, c2, c3), thickness)
-        except:
-            continue
+    a1.grid(color='grey')
+    a1.legend()
+    plt.tight_layout()
 
-def rotatedDottedLine(theta, frame, xi, yi, xf, yf, c1, c2, c3, thickness, segmentLength):
-    centerX = int((xf+xi)/2)
-    centerY = int((yf+yi)/2)
-    lineRadius = int((((xf-xi)**2 + (yf-yi)**2)**(0.5))/2)
-    nXi = int(-lineRadius*np.cos(theta)) + centerX
-    nXf = int(lineRadius*np.cos(theta)) + centerX
-    nYi = int(lineRadius*np.sin(-theta)) + centerY
-    nYf = int(lineRadius*np.sin(theta)) + centerY
-    dottedLine(frame, nXi, nYi, nXf, nYf, c1, c2, c3, thickness, segmentLength)
+    # calculating out the values for each from
+    fps = 30
+    t = np.linspace(start=0, stop=time, num=int(time*fps))
+    frames = len(t)
+    x, y, z = pbd.position(t, omega, u0, v0, x0)
 
-def annotateSideview(img):
-    font = cv2.FONT_HERSHEY_TRIPLEX
+    def init():
+        """Initialization function for the animation."""
+        return puckTop, puckSide
 
-    dpro = 'Side-View'
-    dproLoc = (25, height+borderHeight+25)
-    cv2.putText(img, dpro, dproLoc, font, 1, (255,255,255), 1)
+    def animation_frame(i):
+        """Specific frame of that animation."""
+        xpos = x[i]
+        ypos = y[i]
+        zpos = z[i]
 
-    surf = str(rpm) + ' RPM Parabolic Surface'
-    surfLoc = (width-500, height+borderHeight+25)
-    cv2.putText(img, surf, surfLoc, font, 1, (255,255,255), 1)
+        puckTop.set_data(xpos, ypos)
+        puckSide.set_data(xpos, zpos)
 
-    maxDef = ((omega**2))/(2*9.8)
-    defl = 'Max. Deflection: h = '+ str(round(maxDef,1)) + ' m'
-    deflLoc = (width-500, height+borderHeight+65)
-    cv2.putText(img, defl, deflLoc, font, 1, (255,255,255), 1)
+        return puckTop, puckSide
 
-#def parHeight(x, phi0):
+    animation = animate.FuncAnimation(fig,
+                                      init_func=init,
+                                      func=animation_frame,
+                                      frames=frames,
+                                      blit=True)
+    return animation
 
+def save_animation(animation, name):
+    """Saves animation when called."""
+    animation.save(name, writer='ffmpeg', fps=30, dpi=200)
 
-#def annotateSideview(img, i): # puts diagnostic text info on each frame
-#    font = cv2.FONT_HERSHEY_TRIPLEX
-#
-#    dpro = 'SynthPy: Parabolic Side-View'
-#    dproLoc = (25, 50)
-#    cv2.putText(img, dpro, dproLoc, font, 1, (255, 105, 180), 1)
-#    surf = str(rpm) + ' RPM Parabolic Surface'
-#    surfLoc = (20, 85)
-#    cv2.putText(img, surf, surfLoc, font, 1, (255, 105, 180), 1)
-#
-#    img[25:25+spinlab.shape[0], (width-25)-spinlab.shape[1]:width-25] = spinlab
-#
-#    timestamp = 'Time: ' + str(round((i/fps),1)) + ' s'
-#    tLoc = (width - 225, height-25)
-#    cv2.putText(img, timestamp, tLoc, font, 1, (255, 255, 255), 1)
+# module runs on CLI if run on its own
+if __name__ == "__main__":
+    animation = animate_paraboloid(time, omega, u0, v0, x0, radius)
 
-numFrames = int(movLength * fps)	# calculate number of frames in movie
-phi0 *= -1				# correct angle-measuring convention
-dtheta = rotRate*(6/fps)                    # rotation of camera for each frame (in degrees)
-
-parPoints = np.int32(parabolaPoints()) # cv2 only accepts int32 arrays for polypoints :-|
-parPoints = parPoints.reshape((-1,1,2))
-
-# create map ("dictionary") that allows me to look up the corresponding y-position of any x-point on parabola
-parDict = {}
-for i in range(parPoints.shape[0]):
-    parDict[str(parPoints[i,0,0])] = parPoints[i,0,1]
-
-for i in range(numFrames):
-    frame = np.zeros((height,width,3), np.uint8)
-
-    # Outline of circular table
-    cv2.circle(frame,(width//2, height//2), int(amp), (255,255,255), 2)
-
-    # Place marker at center of table
-    ls = 5 				# line length (in pixels)
-    cv2.line(frame, (width//2+ls, height//2), (width//2-ls, height//2), (255,255,255), 1)
-    cv2.line(frame, (width//2, height//2+ls), (width//2, height//2-ls), (255,255,255), 1)
-
-    dottedLine(frame, int(width/2)-int(amp), int(height/2), int(width/2)+int(amp), int(height/2), 255, 105, 180, 2, 10) # for rotating
-    if rotRate != 0:
-        M = cv2.getRotationMatrix2D((int(width/2), int(height/2)), -i*dtheta*2, 1.0)
-        frame = cv2.warpAffine(frame, M, (width, height))
-
-    # Calculate new position of ball and draw it
-    t = float(i)/fps
-    currentPos = ((width//2)+int(amp*r(t)*np.cos(phi(t))), (height//2)+int(amp*r(t)*np.sin(phi(t))))
-    cv2.circle(frame, currentPos, ballSize, (255,255,255), -1)
-    if rotRate != 0:
-        M = cv2.getRotationMatrix2D((int(width/2), int(height/2)), i*dtheta, 1.0)
-        #dottedLine(frame, int(width/2)-int(amp), int(height/2), int(width/2)+int(amp), int(height/2), 255, 105, 180, 2, 10) # for inertial
-        frame = cv2.warpAffine(frame, M, (width, height))
-        #rotatedDottedLine(-i*dtheta, frame, int(width/2)-int(amp), int(height/2), int(width/2)+int(amp), int(height/2), 255, 105, 180, 2, 10) # for inertial
-    annotate(frame,i, rotRate != 0)
-    #if rotRate != 0:
-        #dottedLine(frame, int(width/2)-int(amp), int(height/2), int(width/2)+int(amp), int(height/2), 55, 255, 90, 2, 10) # for rotating
-    if rotRate == 0:
-        dottedLine(frame, int(width/2)-int(amp), int(height/2), int(width/2)+int(amp), int(height/2), 255, 105, 180, 2, 10) # for inertial
-   #cv2.line(frame, (int(width/2)-int(amp), int(height/2)), (int(width/2)+int(amp), int(height/2)), (255, 105, 180), 2)
-    frame = cv2.resize(frame,(width,height), interpolation = cv2.INTER_CUBIC)
-    if not doParView:
-        video_writer.write(frame)
-    else:
-        fullFrame = np.zeros((fullHeight,width,3), np.uint8)
-
-        # Create parView pad
-        framePar = np.zeros((fullHeight-(height+borderHeight),width,3), np.uint8)
-        if rotRate != 0:
-            cv2.polylines(framePar, [parPoints], 0, (55, 255, 90), 2)
-        else:
-            cv2.polylines(framePar, [parPoints], 0, (255, 105, 180), 2)
-
-        dottedLine(framePar, width//2-int(amp), 30, width//2-int(amp), framePar.shape[0], 255, 255, 255, 2, 10)
-        dottedLine(framePar, width//2+int(amp), 30, width//2+int(amp), framePar.shape[0], 255, 255, 255, 2, 10)
-        #cv2.line(framePar, (width/2-int(amp), 30), (width/2-int(amp), framePar.shape[0]), (255,255,255), 2)
-        #cv2.line(framePar, (width/2+int(amp), 30), (width/2+int(amp), framePar.shape[0]), (255,255,255), 2)
-        if doParViewRot:
-            xpos = (width//2) + int(amp*r(t)*np.cos(phi(t)-((i*dtheta)*(3.14159/180))))
-        else:
-            xpos = (width//2) + int(amp*r(t)*np.cos(phi(t)))
-        currentPos = (xpos, parabola(xpos)-ballSize)
-        cv2.circle(framePar, currentPos, ballSize, (255,255,255), -1)
-        #annotateSideview(framePar, i)
-        #video_writer_par.write(framePar)
-
-        fullFrame[0:height,0:width] = frame
-
-        fullFrame[height+lineWidth:height+(borderHeight-lineWidth),0:width] = 255*np.ones((borderHeight-(2*lineWidth), width, 3), np.uint8)
-        fullFrame[height+borderHeight:fullHeight,0:width] = framePar
-
-
-        dottedLine(fullFrame, width//2-int(amp), height//2, width//2-int(amp), fullFrame.shape[0], 255, 255, 255, 2, 10)
-        dottedLine(fullFrame, width//2+int(amp), height//2, width//2+int(amp), fullFrame.shape[0], 255, 255, 255, 2, 10)
-        #cv2.line(fullFrame, (width/2-int(amp), height/2), (width/2-int(amp), fullFrame.shape[0]), (255,255,255), 2)
-        #cv2.line(fullFrame, (width/2+int(amp), height/2), (width/2+int(amp), fullFrame.shape[0]), (255,255,255), 2)
-        annotateSideview(fullFrame)
-        video_writer.write(fullFrame)
-
-    #if doParView:
-    #    framePar = np.zeros((height,width,3), np.uint8)
-    #    cv2.polylines(framePar, [parPoints], 0, (255,255,255), 2)
-    #    currentPos = ((width/2) + int(amp*r(t)*np.cos(phi(t))), parPoints[i,0,1])
-    #    cv2.circle(framePar, currentPos, ballSize, (255,255,255), -1)
-    #    annotateSideview(framePar, i)
-    #    video_writer_par.write(framePar)
-
-
-
-video_writer.release()
+    if switch == 0:
+        plt.show()
+    elif switch == 1:
+        save_animation(animation, name)
